+++ src/pages/Home/HomePage.jsx
@@ -1,19 +1,20 @@
 import React from "react";
-import HomeHeader from "../components/ui/HomeHeader";
-import HomeContainers from "../components/ui/HomeContainers";
+import HomeHeader from "./ui/HomeHeader";
+import HomeContainers from "./ui/HomeContainers";
+import HomeFooter from "./ui/HomeFooter";
 
 function HomePage() {
     return (
         <>
             <div className="container-fluid">
                 <header id="header-row" className="row">
-                    <HomeHeader />
+                    <HomeHeader/>
                 </header>
                 <main id="home-main" className="row">
-                    <HomeContainers />
+                    <HomeContainers/>
                 </main>
                 <footer id="footer-row" className="row">
-
+                    <HomeFooter/>
                 </footer>
             </div>
         </>
