+++ src/pages/Home/HomePageStyle.scss
@@ -1,0 +1,107 @@
+#home-main{
+    width: 100%;
+    position: fixed;
+    top: 90px;
+    bottom: 90px;
+    //border: 2px solid red;
+}
+
+#home-logo, #navbar{
+    //border: 2px solid yellow;
+    height: 100%;
+    padding: 0;
+    margin: 0;
+}
+
+#header-row{
+    background-color: $bg-transparent;
+    height: 90px;
+    width: 100%;
+    position: fixed;
+    top: 0;
+    padding: 0px;
+    padding: 0;
+    //border: 2px solid red;
+    #home-logo{
+        padding: 10px 5px 5px 0px;
+        //border: 1px solid blue;
+        #jam-logo{
+            display: flex;
+            justify-content: flex-start;
+            //border: 1px solid blue;
+        }
+        #spotify-logo{
+            display: flex;
+            justify-content: flex-start;
+            position: relative;
+            top: -32px;
+            left: 150px;
+            //border: 2px solid red;
+            #with-text{
+                position: relative;
+                top: 10px;
+                font-size: 0.75em;
+            }
+        }
+    }
+    #navbar{
+        //border: 2px solid red;
+        display: flex;
+        justify-content: flex-end;
+        flex-wrap: wrap;
+        align-items: center;
+        padding: 0;
+        margin: 0;
+        ul{
+            //border: 2px solid red;
+            display: flex;
+            justify-content: flex-end;
+            width: 500px;
+            gap: 10px;
+            align-items: center;
+            padding: 0;
+            margin: 0 25px;
+            list-style: none;
+            font-weight: bold;
+            font-size: 1.1rem;
+            a{
+                color: white;
+                text-decoration: none;
+            }
+            a:hover{
+                text-decoration: underline;
+            }
+            a:visited{
+                color: white;
+                text-decoration: none;
+            }
+            #about-button{
+                color: white;
+                text-decoration: none;
+                font-weight: bold;
+            }
+    
+        }
+    }
+}
+
+.popover{
+    background-color: lightgray;
+}
+
+//--------------------------------------------------------------
+
+#playlists-box{
+    //border: 2px solid red;
+    border-radius: 20px;
+    background-color: $bg-transparent;
+    padding: 0;
+    margin: 15px 7.5px 15px 15px;
+}
+#search-box{
+    //border: 2px solid red;
+    background-color: $bg-transparent;
+    border-radius: 20px;
+    padding: 0;
+    margin: 15px 15px 15px 7.5px;
+}
