+++ src/pages/Home/ui/HomeContainers.jsx
@@ -3,7 +3,7 @@ import React from "react";
 function HomeContainers(){
     return(
         <>
-            <section id="playlists-box" className="col-4">
+            <section id="playlists-box" className="col">
 
             </section>
             <section id="search-box" className="col">
