+++ src/pages/Home/ui/HomeFooter.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function HomeFooter(){
+    return(
+        <>
+            <section id="" className="col">
+
+            </section>
+        </>
+    )
+}
+
+export default HomeFooter;
